api/requirements.txt
uvicorn >= 0.11.3
fastapi >= 0.54.1
requests >= 2.23.0
pydantic >= 1.4
pydantic[email] >= 1.4
gunicorn >= 20.0.4
psycopg2 >= 2.8.5
sqlalchemy >= 1.3.16
email-validator >= 1.0.5
google-api-python-client
google-auth 
google-auth-oauthlib 
google-auth-httplib2
oauth2client
